+++ src/core/const/props.ts
@@ -6,11 +6,6 @@
  * https://github.com/V4Fire/Core/blob/master/LICENSE
  */
 
-/**
- * [[include:core/const/README.md]]
- * @packageDocumentation
- */
-
 export const defProp = Object.freeze({
 	configurable: true,
 	enumerable: true,